+++ lib/basic/entity.dart
@@ -8,6 +8,9 @@ class Entity implements EntityClass{
   static Function Player = ({Range distance,Range level, Gamemode gamemode, List<Location> area, String name, Range horizontalRotation, Range verticalRotation}){
     return Entity(selector: 'p',distance:distance,level: level,area: area,gamemode: gamemode,name: name,horizontalRotation: horizontalRotation,verticalRotation: verticalRotation);
   };
+  static Function PlayerName = (String name){
+    return Entity(playerName: name);
+  };
   static Function All = ({Range distance,int limit,Range level, Gamemode gamemode, List<Location> area, String name, Range horizontalRotation, Range verticalRotation}){
     return Entity(selector: 'a',limit:limit,distance:distance,level: level,area: area,gamemode: gamemode,name: name,horizontalRotation: horizontalRotation,verticalRotation: verticalRotation);
   };
@@ -28,7 +31,8 @@ class Entity implements EntityClass{
       Gamemode gamemode, 
       String name, 
       Range horizontalRotation, 
-      Range verticalRotation}){
+      Range verticalRotation,
+      this.playerName}){
 
     if(distance != null) arguments['distance'] = distance.toString();
     if(level != null) arguments['level'] = level.toString();
@@ -52,13 +56,16 @@ class Entity implements EntityClass{
   Map arguments = {
     
   };
+  String playerName;
   Entity sort(Sort sort){
     arguments['sort'] = sort.toString().split('.').last;
     return this;
   }
   @override
   String toString(){
+    if(playerName != null && playerName.isNotEmpty) return playerName;
     String ret = "@" + selector;
+    
     if(arguments.length > 0){
       ret += '[';
       arguments.keys.forEach((key){
@@ -110,106 +117,106 @@ enum Sort{
 }
 class EntityType {
   // throwable
-  static EntityType armor_stand = EntityType('armor_stand');
-  static EntityType area_effect_cloud = EntityType('area_effect_cloud');
-  static EntityType player = EntityType('player');
-  static EntityType boat = EntityType('boat');
-  static EntityType minecart	 = EntityType('minecart');
-  static EntityType chest_minecart	 = EntityType('chest_minecart');
-  static EntityType furnace_minecart	 = EntityType('furnace_minecart');
-  static EntityType tnt_minecart	 = EntityType('tnt_minecart');
-  static EntityType hopper_minecart	 = EntityType('hopper_minecart');
-  static EntityType spawner_minecart	 = EntityType('spawner_minecart');
-  static EntityType item	 = EntityType('item');
-  static EntityType experience_orb	 = EntityType('experience_orb');
-  static EntityType experience_bottle	 = EntityType('experience_bottle');
-  static EntityType arrow	 = EntityType('arrow');
-  static EntityType spectral_arrow	 = EntityType('spectral_arrow');
-  static EntityType trident	 = EntityType('trident');
-  static EntityType snowball	 = EntityType('snowball');
-  static EntityType egg	 = EntityType('egg');
-  static EntityType llama_spit	 = EntityType('llama_spit');
-  static EntityType ender_pearl	 = EntityType('ender_pearl');
-  static EntityType eye_of_ender	 = EntityType('eye_of_ender');
-  static EntityType fireworks_rocket	 = EntityType('fireworks_rocket');
-  static EntityType tnt	 = EntityType('tnt');
-  static EntityType falling_block	 = EntityType('falling_block');
-  static EntityType fishing_bobber	 = EntityType('fishing_bobber');
-  static EntityType leash_knot	 = EntityType('leash_knot');
-  static EntityType painting	 = EntityType('painting');
-  static EntityType item_frame	 = EntityType('item_frame');
-  static EntityType fireball	 = EntityType('fireball');
-  static EntityType shulker_bullet	 = EntityType('shulker_bullet');
-  static EntityType end_crystal	 = EntityType('end_crystal');
-  static EntityType evoker_fangs	 = EntityType('evoker_fangs');
-  static EntityType potion = EntityType('potion');
+  static const EntityType armor_stand = const EntityType('armor_stand');
+  static const EntityType area_effect_cloud = const EntityType('area_effect_cloud');
+  static const EntityType player = const EntityType('player');
+  static const EntityType boat = const EntityType('boat');
+  static const EntityType minecart	 = const EntityType('minecart');
+  static const EntityType chest_minecart	 = const EntityType('chest_minecart');
+  static const EntityType furnace_minecart	 = const EntityType('furnace_minecart');
+  static const EntityType tnt_minecart	 = const EntityType('tnt_minecart');
+  static const EntityType hopper_minecart	 = const EntityType('hopper_minecart');
+  static const EntityType spawner_minecart	 = const EntityType('spawner_minecart');
+  static const EntityType item	 = const EntityType('item');
+  static const EntityType experience_orb	 = const EntityType('experience_orb');
+  static const EntityType experience_bottle	 = const EntityType('experience_bottle');
+  static const EntityType arrow	 = const EntityType('arrow');
+  static const EntityType spectral_arrow	 = const EntityType('spectral_arrow');
+  static const EntityType trident	 = const EntityType('trident');
+  static const EntityType snowball	 = const EntityType('snowball');
+  static const EntityType egg	 = const EntityType('egg');
+  static const EntityType llama_spit	 = const EntityType('llama_spit');
+  static const EntityType ender_pearl	 = const EntityType('ender_pearl');
+  static const EntityType eye_of_ender	 = const EntityType('eye_of_ender');
+  static const EntityType fireworks_rocket	 = const EntityType('fireworks_rocket');
+  static const EntityType tnt	 = const EntityType('tnt');
+  static const EntityType falling_block	 = const EntityType('falling_block');
+  static const EntityType fishing_bobber	 = const EntityType('fishing_bobber');
+  static const EntityType leash_knot	 = const EntityType('leash_knot');
+  static const EntityType painting	 = const EntityType('painting');
+  static const EntityType item_frame	 = const EntityType('item_frame');
+  static const EntityType fireball	 = const EntityType('fireball');
+  static const EntityType shulker_bullet	 = const EntityType('shulker_bullet');
+  static const EntityType end_crystal	 = const EntityType('end_crystal');
+  static const EntityType evoker_fangs	 = const EntityType('evoker_fangs');
+  static const EntityType potion = const EntityType('potion');
 
   // hostile
-  static EntityType elder_guardian	 = EntityType('elder_guardian');
-  static EntityType wither_skeleton	 = EntityType('wither_skeleton');
-  static EntityType stray	 = EntityType('stray');
-  static EntityType husk	 = EntityType('husk');
-  static EntityType zombie_villager	 = EntityType('zombie_villager');
-  static EntityType vex	 = EntityType('vex');
-  static EntityType vindicator	 = EntityType('vindicator');
-  static EntityType illager_beast	 = EntityType('illager_beast');
-  static EntityType illusioner	 = EntityType('illusioner');
-  static EntityType pillager	 = EntityType('pillager');
-  static EntityType creeper	 = EntityType('creeper');
-  static EntityType skeleton	 = EntityType('skeleton');
-  static EntityType spider	 = EntityType('spider');
-  static EntityType giant	 = EntityType('giant');
-  static EntityType zombie	 = EntityType('zombie');
-  static EntityType slime	 = EntityType('slime');
-  static EntityType ghast	 = EntityType('ghast');
-  static EntityType zombie_pigman	 = EntityType('zombie_pigman');
-  static EntityType enderman	 = EntityType('enderman');
-  static EntityType cave_spider	 = EntityType('cave_spider');
-  static EntityType silverfish	 = EntityType('silverfish');
-  static EntityType blaze	 = EntityType('blaze');
-  static EntityType magma_cube = EntityType('magma_cube');
-  static EntityType ender_dragon = EntityType('ender_dragon');
-  static EntityType wither = EntityType('wither');
-  static EntityType wither_skull = EntityType('wither_skull');
-  static EntityType witch = EntityType('witch');
-  static EntityType endermite = EntityType('endermite');
-  static EntityType guardian = EntityType('guardian');
-  static EntityType shulker = EntityType('shulker');
-  static EntityType dragon_fireball = EntityType('dragon_fireball');
-  static EntityType drowned = EntityType('drowned');
+  static const EntityType elder_guardian	 = const EntityType('elder_guardian');
+  static const EntityType wither_skeleton	 = const EntityType('wither_skeleton');
+  static const EntityType stray	 = const EntityType('stray');
+  static const EntityType husk	 = const EntityType('husk');
+  static const EntityType zombie_villager	 = const EntityType('zombie_villager');
+  static const EntityType vex	 = const EntityType('vex');
+  static const EntityType vindicator	 = const EntityType('vindicator');
+  static const EntityType illager_beast	 = const EntityType('illager_beast');
+  static const EntityType illusioner	 = const EntityType('illusioner');
+  static const EntityType pillager	 = const EntityType('pillager');
+  static const EntityType creeper	 = const EntityType('creeper');
+  static const EntityType skeleton	 = const EntityType('skeleton');
+  static const EntityType spider	 = const EntityType('spider');
+  static const EntityType giant	 = const EntityType('giant');
+  static const EntityType zombie	 = const EntityType('zombie');
+  static const EntityType slime	 = const EntityType('slime');
+  static const EntityType ghast	 = const EntityType('ghast');
+  static const EntityType zombie_pigman	 = const EntityType('zombie_pigman');
+  static const EntityType enderman	 = const EntityType('enderman');
+  static const EntityType cave_spider	 = const EntityType('cave_spider');
+  static const EntityType silverfish	 = const EntityType('silverfish');
+  static const EntityType blaze	 = const EntityType('blaze');
+  static const EntityType magma_cube = const EntityType('magma_cube');
+  static const EntityType ender_dragon = const EntityType('ender_dragon');
+  static const EntityType wither = const EntityType('wither');
+  static const EntityType wither_skull = const EntityType('wither_skull');
+  static const EntityType witch = const EntityType('witch');
+  static const EntityType endermite = const EntityType('endermite');
+  static const EntityType guardian = const EntityType('guardian');
+  static const EntityType shulker = const EntityType('shulker');
+  static const EntityType dragon_fireball = const EntityType('dragon_fireball');
+  static const EntityType drowned = const EntityType('drowned');
 
   // passive
-  static EntityType skeleton_horse = EntityType('skeleton_horse');
-  static EntityType zombie_horse = EntityType('zombie_horse');
-  static EntityType donkey = EntityType('donkey');
-  static EntityType mule = EntityType('mule');
-  static EntityType bat = EntityType('bat');
-  static EntityType pig = EntityType('pig');
-  static EntityType sheep = EntityType('sheep');
-  static EntityType cow = EntityType('cow');
-  static EntityType chicken = EntityType('chicken');
-  static EntityType squid = EntityType('squid');
-  static EntityType wolf = EntityType('wolf');
-  static EntityType mooshroom = EntityType('mooshroom');
-  static EntityType snowman = EntityType('snowman');
-  static EntityType ocelot = EntityType('ocelot');
-  static EntityType iron_golem = EntityType('iron_golem');
-  static EntityType horse = EntityType('horse');
-  static EntityType rabbit = EntityType('rabbit');
-  static EntityType polar_bear = EntityType('polar_bear');
-  static EntityType llama = EntityType('llama');
-  static EntityType parrot = EntityType('parrot');
-  static EntityType villager = EntityType('villager');
-  static EntityType cod = EntityType('cod');
-  static EntityType salmon = EntityType('salmon');
-  static EntityType tropical_fish = EntityType('tropical_fish');
-  static EntityType turtle = EntityType('turtle');
-  static EntityType dolphin = EntityType('dolphin');
-  static EntityType panda = EntityType('panda');
+  static const EntityType skeleton_horse = const EntityType('skeleton_horse');
+  static const EntityType zombie_horse = const EntityType('zombie_horse');
+  static const EntityType donkey = const EntityType('donkey');
+  static const EntityType mule = const EntityType('mule');
+  static const EntityType bat = const EntityType('bat');
+  static const EntityType pig = const EntityType('pig');
+  static const EntityType sheep = const EntityType('sheep');
+  static const EntityType cow = const EntityType('cow');
+  static const EntityType chicken = const EntityType('chicken');
+  static const EntityType squid = const EntityType('squid');
+  static const EntityType wolf = const EntityType('wolf');
+  static const EntityType mooshroom = const EntityType('mooshroom');
+  static const EntityType snowman = const EntityType('snowman');
+  static const EntityType ocelot = const EntityType('ocelot');
+  static const EntityType iron_golem = const EntityType('iron_golem');
+  static const EntityType horse = const EntityType('horse');
+  static const EntityType rabbit = const EntityType('rabbit');
+  static const EntityType polar_bear = const EntityType('polar_bear');
+  static const EntityType llama = const EntityType('llama');
+  static const EntityType parrot = const EntityType('parrot');
+  static const EntityType villager = const EntityType('villager');
+  static const EntityType cod = const EntityType('cod');
+  static const EntityType salmon = const EntityType('salmon');
+  static const EntityType tropical_fish = const EntityType('tropical_fish');
+  static const EntityType turtle = const EntityType('turtle');
+  static const EntityType dolphin = const EntityType('dolphin');
+  static const EntityType panda = const EntityType('panda');
 
 
-  String type;
-  EntityType(this.type);
+  final String type;
+  const EntityType(this.type);
   @override
     String toString() {
       return type;